requirements.txt
Click==7.0
cycler==0.10.0
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
imutils==0.5.2
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
pyparsing==2.4.0
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
torch==1.0.1.post2
torchvision==0.2.2.post3
